Finally put to bed the stupid duplicate company name folders

+++ Tools/SFCKMO2Mode/updateRepositoryFromMO2.ps1
@@ -10,8 +10,12 @@ if (!$Global:SharedConfigurationLoaded) {
 
 # Need to copy the source scripts to the Scripts Source folder so SFCK can use them
 Write-Host -ForegroundColor Green "Copying the source scripts to the Scripts Source folder so SFCK can use them"
-Copy-Item -Recurse -Force -Path "$ENV:MODULE_SCRIPTS_SOURCE_PATH\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceModule\**" -Destination ".\Source\Papyrus\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceModule"
-Copy-Item -Recurse -Force -Path "$ENV:MODULE_SCRIPTS_SOURCE_PATH\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceSharedLibrary\**" -Destination ".\Source\Papyrus\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceSharedLibrary"
+If ([System.IO.Directory]::Exists("$ENV:MODULE_SCRIPTS_PATH\$Global:ScriptingNamespaceCompany")) {
+  Copy-Item -Recurse -Force -Path "$ENV:MODULE_SCRIPTS_SOURCE_PATH\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceModule\**" -Destination ".\Source\Papyrus\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceModule"
+  Copy-Item -Recurse -Force -Path "$ENV:MODULE_SCRIPTS_SOURCE_PATH\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceSharedLibrary\**" -Destination ".\Source\Papyrus\$Global:ScriptingNamespaceCompany\$Global:ScriptingNamespaceSharedLibrary"
+} Else {
+  Write-Host -ForegroundColor Red "WARNING: No scripting support detected so no scripts to sync."
+}
 
 # Need to copy the ESM/ESP/ESL files from the Game Data folder to update the repository verison for commit
 Write-Host -ForegroundColor Green "Copying the ESM/ESP/ESL files from the Game Data folder to update the repository verison for commit"
